# apps/functions/openapi.yaml
openapi: 3.0.0
info:
  title: TODO App API
  version: 1.0.0
  description: API REST para aplicación TODO con autenticación JWT y gestión de tareas por usuario
  contact:
    name: Mikel Ortega
  license:
    name: MIT

servers:
  - url: http://localhost:5001/demo-atom-challenge/us-central1/api/v1
    description: Development (Emulators)
  - url: https://us-central1-atom-challenge-mikel.cloudfunctions.net/api/v1
    description: Production

tags:
  - name: Health
    description: Health check endpoints
  - name: Users
    description: Authentication endpoints (public)
  - name: Tasks
    description: Task management (protected)

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: OK

  /users/login:
    post:
      tags: [Users]
      summary: Login por email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Login exitoso
        '404':
          description: Usuario no encontrado

  /users:
    post:
      tags: [Users]
      summary: Crear usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '201':
          description: Usuario creado
        '409':
          description: Usuario ya existe

  /tasks:
    get:
      tags: [Tasks]
      summary: Listar tareas
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de tareas
        '401':
          description: No autenticado
    post:
      tags: [Tasks]
      summary: Crear tarea
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 80
                description:
                  type: string
                  maxLength: 200
              required: [title]
      responses:
        '201':
          description: Tarea creada
        '401':
          description: No autenticado

  /tasks/{id}:
    patch:
      tags: [Tasks]
      summary: Actualizar tarea
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tarea actualizada
        '403':
          description: Sin permiso
        '404':
          description: No encontrada
    delete:
      tags: [Tasks]
      summary: Eliminar tarea
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminada
        '403':
          description: Sin permiso

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT