# apps/functions/src/app.ts
/**
 * Express Application
 * API REST para TODO App
 */

import express, { Request, Response, NextFunction } from 'express';
import cors from 'cors';
import helmet from 'helmet';
import swaggerUi from 'swagger-ui-express';
import * as usersController from './api/users.controller';
import * as tasksController from './api/tasks.controller';
import { authMiddleware } from './middleware/auth';
import { swaggerSpec } from './docs/openapi';

// Inicializar app
const app = express();

// Middlewares de seguridad
app.use(helmet());

// CORS - permitir orígenes específicos
const allowedOrigins = process.env.ALLOWED_ORIGINS
  ? process.env.ALLOWED_ORIGINS.split(',')
  : ['http://localhost:4200', 'http://localhost:5002'];

app.use(
  cors({
    origin: allowedOrigins,
    credentials: true,
  })
);

// Parse JSON
app.use(express.json());

// API Documentation (Swagger UI) - Public
app.use('/v1/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));

/**
 * @openapi
 * /health:
 *   get:
 *     tags:
 *       - Health
 *     summary: Health check
 *     description: Verifica que la API esté funcionando correctamente
 *     responses:
 *       200:
 *         description: API funcionando correctamente
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 status:
 *                   type: string
 *                   example: ok
 *                 message:
 *                   type: string
 *                   example: TODO API is running
 *                 timestamp:
 *                   type: string
 *                   format: date-time
 *                   example: 2025-10-21T18:30:00.000Z
 *                 version:
 *                   type: string
 *                   example: 1.0.0
 */
app.get('/v1/health', (_req: Request, res: Response) => {
  res.status(200).json({
    status: 'ok',
    message: 'TODO API is running',
    timestamp: new Date().toISOString(),
    version: '1.0.0',
  });
});

// Users routes (public - no auth required)
app.post('/v1/users/login', usersController.login);
app.post('/v1/users', usersController.createUser);

// Auth middleware - todas las rutas después requieren autenticación
app.use('/v1', authMiddleware);

// Tasks routes (protected - auth required)
app.get('/v1/tasks', tasksController.list);
app.post('/v1/tasks', tasksController.create);
app.patch('/v1/tasks/:id', tasksController.update);
app.delete('/v1/tasks/:id', tasksController.remove);

// 404 handler
app.use((_req: Request, res: Response) => {
  res.status(404).json({
    error: 'Not Found',
    message: 'The requested endpoint does not exist',
  });
});

// Error handler
app.use((err: Error, _req: Request, res: Response, _next: NextFunction) => {
  console.error('Error:', err);
  res.status(500).json({
    error: 'Internal Server Error',
    message: process.env.NODE_ENV === 'development' ? err.message : 'Something went wrong',
  });
});

export default app;